workers/requirements.txt
celery[redis]==5.4.0
redis[hiredis]==5.0.7
psycopg2-binary==2.9.9
httpx==0.27.0
python-dotenv==1.0.1
